% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpc.R
\name{computeCPVPC}
\alias{computeCPVPC}
\title{Compute the Variance Partition Coefficient (VPC) for Compound Poisson Model}
\usage{
computeCPVPC(para, group)
}
\arguments{
\item{para}{A matrix or vector of model parameters for the Compound Poisson model.}

\item{group}{a scalar 0, 1 for binary group or some float for continuous variable}
}
\value{
A matrix with computed VPC values for the Compound Poisson model. The number of rows matches the number of rows of the input \code{para}.
}
\description{
Computes the VPC for a given set of parameters for the Compound Poisson model.
It checks the dimension of the input parameters to determine the type of computation (with slope or without).
}
\examples{
# Sample use of computeCPVPC function
sample_params <- matrix(rnorm(25), ncol=5)
computeCPVPC(sample_params,0)

}
\seealso{
\code{\link{computeVPC}}
}
