% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpc.R
\name{computeVPC}
\alias{computeVPC}
\title{Compute the Variance Partition Coefficient (VPC)}
\usage{
computeVPC(para, covariate, type)
}
\arguments{
\item{para}{A matrix or vector of model parameters.}

\item{covariate}{A numeric vector of covariateiate values corresponding to the rows of \code{para}.}

\item{type}{Character string specifying the type of model ('NB' for Negative Binomial or 'CP' for Compound Poisson).}
}
\value{
A matrix with computed VPC values. The number of rows matches the number of rows of the input \code{para}.
}
\description{
An internal function that computes the VPC for given parameters and a covariateiate.
This function can handle both single parameter vectors and matrices.
}
\keyword{internal}
