% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{extractParams}
\alias{extractParams}
\title{Extract Model Parameters}
\usage{
extractParams(model, type, slope = FALSE)
}
\arguments{
\item{model}{A fitted \code{glmmTMB} model object.}

\item{type}{A character string specifying the type of the model, either "NB" for negative binomial
or "CP" for compound Poisson.}

\item{slope}{A logical indicating whether the model includes a slope for the random effect.
Default is FALSE.}
}
\value{
A numeric vector of extracted model parameters.
}
\description{
Internal function to extract fixed effects, variance components, and other parameters
from a glmmTMB model.
}
\keyword{internal}
