% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extractStrainFactor}
\alias{extractStrainFactor}
\title{Extract a Specific Part of Column Names and Convert to Factor}
\usage{
extractStrainFactor(countMatrix, partIndex, splitSymbol = "_")
}
\arguments{
\item{countMatrix}{A matrix or data frame with column names.}

\item{partIndex}{An integer specifying which part of the split column names to return.}

\item{splitSymbol}{A string specifying the delimiter to use when splitting column names. Default is "_".}
}
\value{
A factor representing the extracted part of the column names.
}
\description{
This function takes a matrix, splits its column names based on a specified delimiter,
and then returns a factor representing the part of the split name corresponding to a specified position.
}
\examples{
testMatrix <- matrix(1:6, ncol=3)
colnames(testMatrix) <- c("strainA_part1_result", "strainB_part2_result", "strainC_part3_result")
extractStrainFactor(testMatrix, 1)
}
