% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fitGeneralizedModel}
\alias{fitGeneralizedModel}
\title{Fit Generalized Linear Mixed-Effects Model to Count Matrix}
\usage{
fitGeneralizedModel(countMatrix, X, type, slope = FALSE, parallel = 1)
}
\arguments{
\item{countMatrix}{A matrix where rows represent features and columns represent samples.}

\item{X}{A numeric vector representing the covariate.}

\item{type}{A character string specifying the type of model to fit. Either "NB" (negative binomial) or "CP" (tweedie).}

\item{slope}{Logical. If TRUE, the random effect will also have a slope component with the covariate. Default is FALSE.}

\item{parallel}{integer Set number of OpenMP threads to evaluate the negative log-likelihood in parallel}
}
\value{
A matrix with rows corresponding to features in \code{countMatrix} and columns representing the model parameters. Row names of the returned matrix match the row names of the input \code{countMatrix}.
}
\description{
This function fits a generalized linear mixed-effects model to each row of a provided count matrix using the \code{glmmTMB} package.
}
\seealso{
\code{\link[glmmTMB]{glmmTMB}}
}
\keyword{internal}
