% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fitModelCommon}
\alias{fitModelCommon}
\title{Fit Generalized Model to Each Row of Count Matrix}
\usage{
fitModelCommon(countMatrix, X, slope, type, parallel, returnWarnings = FALSE)
}
\arguments{
\item{countMatrix}{A matrix where rows represent features and columns represent samples.}

\item{X}{A numeric vector representing the covariate.}

\item{slope}{Logical. Determines whether the random effect should include a slope component with the covariate. Default is FALSE.}

\item{type}{Character string indicating the model type: "NB" for Negative Binomial or "CP" for Compound Poisson (Tweedie).}

\item{parallel}{Numeric value indicating the number of OpenMP threads to use for parallel processing.}

\item{returnWarnings}{Logical. If TRUE, the function will return any warnings that were generated during the model fitting process.}
}
\value{
A list where each element corresponds to the result of fitting the model
to a row of the countMatrix. If \code{returnWarnings} is TRUE, this list will contain
any warnings that were generated. Otherwise, it will contain model parameters or NA values for failed fits.
}
\description{
This function fits a generalized model (either Negative Binomial or Tweedie)
to each row of a given count matrix. It returns the model results and
optionally captures any warnings that are generated during the fitting process.
}
\examples{
# This is a placeholder. Update with a relevant example:
# countMatrix <- your_example_data
# X <- your_example_data
# fitModelCommon(countMatrix, X, slope=TRUE, type="NB", parallel=2, returnWarnings=TRUE)

}
\seealso{
\code{\link{captureModelWarnings}}, \code{\link{fitGeneralizedModel}}
}
