% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fitNBmodel}
\alias{fitNBmodel}
\title{Fit a Negative Binomial Model to a Count Matrix}
\usage{
fitNBmodel(countMatrix, X, slope = FALSE, parallel = 1, reportWarning = FALSE)
}
\arguments{
\item{countMatrix}{A matrix where rows represent features and columns represent samples.}

\item{X}{A numeric vector representing the covariate.}

\item{slope}{Logical. If TRUE, the random effect will also have a slope component
with the covariate. Default is FALSE.}

\item{parallel}{Set number of OpenMP threads to evaluate the negative log-likelihood in parallel.}

\item{reportWarning}{Logical. If TRUE then the model is fit twice to obtain the warning logs.}
}
\value{
A list containing the model parameters and, if \code{reportWarning} is TRUE,
the warnings generated during model fitting.
}
\description{
Fits a Negative Binomial model to each row of a provided count matrix.
}
\seealso{
\code{\link{fitGeneralizedModel}}, \code{\link{captureModelWarnings}}
}
