% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generateCovariate}
\alias{generateCovariate}
\title{Covariate Generation from Specified Distribution}
\usage{
generateCovariate(strainSizes, distribution = "rgen01", ...)
}
\arguments{
\item{strainSizes}{A numeric vector of integers.
Each element denotes the number of samples within a corresponding strain.
The length of this vector indicates the total number of strains.}

\item{distribution}{A function representing the desired distribution from
which the covariate values will be generated. By default, the function
generates a binary covariate. For functions in \code{R} use the generating functions
e.g \code{rnorm}, \code{rpois}, \code{rgamma} e.t.c}

\item{...}{Optional named parameters specific to the provided \code{distribution}
function}
}
\value{
A numeric vector representing the generated covariate from the
specified distribution.
}
\description{
Covariate Generation from Specified Distribution
}
\examples{
strainSizes <- c(10, 15, 20)
generateCovariate(strainSizes)
generateCovariate(strainSizes, distribution = "rnorm")
generateCovariate(strainSizes, distribution = "rgamma", shape=2)
}
