% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generateParamPlots}
\alias{generateParamPlots}
\title{Generate Histogram Plots for Parameters in a Matrix}
\usage{
generateParamPlots(paramMatrix, col = "lightblue", border = "black")
}
\arguments{
\item{paramMatrix}{A matrix with numeric data.}

\item{col}{Color of the histogram bars. Default is "lightblue".}

\item{border}{Border color of the histogram bars. Default is "black".}
}
\description{
Generate Histogram Plots for Parameters in a Matrix
}
\examples{
set.seed(123) # Setting a seed for reproducibility
sampleMatrix <- matrix(rnorm(300), ncol=3)
colnames(sampleMatrix) <- c("Parameter_1", "Parameter_2", "Parameter_3")

# Use the generateParamPlots function to create histograms
generateParamPlots(sampleMatrix)
}
