% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateCount.R
\name{generateReadMatrix}
\alias{generateReadMatrix}
\title{Generate a Matrix of Read Counts for Multiple Features}
\usage{
generateReadMatrix(
  strainSizes,
  alphas,
  betas,
  sigma2s,
  phis,
  X,
  method = c("NB", "CP"),
  ps = NULL
)
}
\arguments{
\item{strainSizes}{A numeric vector specifying the size of each strain.}

\item{alphas}{A numeric vector of intercepts.}

\item{betas}{A numeric vector of regression coefficients.}

\item{sigma2s}{A list of variance-covariance matrices for random effects, one for each feature.}

\item{phis}{A numeric vector of dispersion parameters for the count distribution.}

\item{X}{A numeric vector or matrix of covariates.}

\item{method}{A character string. It can be either "NB" (Negative Binomial) or "CP" (some other method). Default is "NB".}

\item{ps}{A numeric vector. Only used if method is "CP". Default is NULL.}
}
\value{
A matrix of read counts, where each row corresponds to a feature and each column
corresponds to a sample/strain.
}
\description{
This function produces a matrix of read counts for multiple features
based on the provided parameters and method specified (\code{NB} or \code{CP}).
It is designed to be an internal helper function.
}
\keyword{internal}
