% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateCount.R
\name{generateReads}
\alias{generateReads}
\title{Generate Counts from Given Parameters and Model}
\usage{
generateReads(strainSizes, alpha, beta, sigma, X, phi, countFun, ...)
}
\arguments{
\item{strainSizes}{A numeric vector specifying the size of each strain.}

\item{alpha}{A numeric value representing the intercept term.}

\item{beta}{A numeric value for the regression coefficient.}

\item{sigma}{A numeric matrix or value for the random effects covariance structure.}

\item{X}{A numeric vector or matrix of covariates.}

\item{phi}{A numeric dispersion parameter passed to the \code{countFun}.}

\item{countFun}{A function for generating count data. This function should accept parameters \code{n}, \code{mu}, and \code{phi} at minimum.}

\item{...}{Additional arguments passed to the \code{countFun}.}
}
\value{
A matrix with counts. Each column corresponds to a different strain-sample and is named accordingly.
}
\description{
This function produces counts based on provided parameters, a mean function, and a count function.
It uses the \code{generateMeans} function to determine the mean structure and then applies a count-generating function.
}
\keyword{internal}
