% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateCount.R
\name{getCPReads}
\alias{getCPReads}
\title{Generate Compound Poisson (Tweedie) Reads}
\usage{
getCPReads(strainSizes, alpha, beta, sigma, p, phi, X)
}
\arguments{
\item{strainSizes}{A numeric vector specifying the size of each strain.}

\item{alpha}{A numeric specifying the intercept.}

\item{beta}{A numeric specifying the regression coefficient.}

\item{sigma}{A numeric or matrix specifying the variance (or variance-covariance matrix) for random effects.}

\item{p}{A numeric specifying the power parameter for the Tweedie distribution.}

\item{phi}{A numeric specifying the dispersion parameter for the Tweedie distribution.}

\item{X}{A numeric vector or matrix of covariates.}
}
\value{
A numeric vector of Compound Poisson (Tweedie) read counts.
}
\description{
This function produces a vector of read counts generated from a Compound Poisson (Tweedie) distribution
based on the provided parameters.
}
\seealso{
\code{\link[tweedie]{rtweedie}}
}
