% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateCount.R
\name{getNBReadMatrix}
\alias{getNBReadMatrix}
\title{Generate Read Matrix Using Negative Binomial Distribution}
\usage{
getNBReadMatrix(strainSizes, alphas, betas, sigma2s, phis, X)
}
\arguments{
\item{strainSizes}{A numeric vector specifying the size of each strain.}

\item{alphas}{A numeric vector of intercepts.}

\item{betas}{A numeric vector of regression coefficients.}

\item{sigma2s}{A list of numeric values or matrices specifying the variances (or variance-covariance matrices)
for random effects for each gene.}

\item{phis}{A numeric vector specifying the dispersion parameters for the Negative Binomial distribution.}

\item{X}{A numeric vector or matrix of covariates.}
}
\value{
A numeric matrix of Negative Binomial read counts where each row represents a gene
and each column corresponds to a sample.
}
\description{
This function produces a matrix of read counts generated from a Negative Binomial distribution
based on the provided parameters.
}
\seealso{
\code{\link[MASS]{rnegbin}}
}
