% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getPartOfColNames}
\alias{getPartOfColNames}
\title{Extract Specific Parts of Column Names Based on a Delimiter}
\usage{
getPartOfColNames(countMatrix, partIndex, splitSymbol = "_")
}
\arguments{
\item{countMatrix}{A matrix containing the data. This matrix should have column names.}

\item{partIndex}{An integer specifying which part of the split column names to return.
For example, if \code{partIndex} is 2 and the column name is "A_B_C", the function will return "B".}

\item{splitSymbol}{A string specifying the delimiter to use when splitting column names.
Default is "_".}
}
\value{
A character vector containing the extracted parts of the column names.
}
\description{
This function takes a matrix, splits its column names based on a specified delimiter,
and then returns the part of the split name corresponding to a specified position.
}
\examples{
testMatrix <- matrix(1:6, ncol=3)
colnames(testMatrix) <- c("name#1#test", "name#2", "name#3#extra#info")
getPartOfColNames(testMatrix, 2, splitSymbol="#")
}
