% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeritSeq2-package.R
\docType{package}
\name{HeritSeq2-package}
\alias{HeritSeq2}
\alias{HeritSeq2-package}
\title{HeritSeq2: Estimating Heritability from High-Throughput Sequencing Data using Generalized Linear Mixed Models}
\description{
Provides a comprehensive toolkit to apply Generalized Linear Mixed Models (GLMM) for estimating the heritability of high-throughput sequencing data. This package aims to streamline the analysis of large-scale sequencing datasets, enabling researchers to derive meaningful genetic insights from complex genomic data.
}
\author{
\strong{Maintainer}: Julius Olaifa \email{julius.olaifa@okstate.edu} (\href{https://orcid.org/0000-0002-3864-099X}{ORCID})

}
\keyword{internal}
