% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fitNBmodel}
\alias{fitNBmodel}
\title{Fit Negative Binomial Generalized Linear Mixed-Effects Model to Count Matrix}
\usage{
fitNBmodel(countMatrix, X, slope = FALSE)
}
\arguments{
\item{countMatrix}{A matrix where rows represent features and columns represent samples.}

\item{X}{A numeric vector representing the covariate.}

\item{slope}{Logical. If TRUE, the random effect will also have a slope component with the covariate. Default is FALSE.}
}
\value{
A matrix with rows corresponding to features in \code{countMatrix} and columns representing the model parameters. Row names of the returned matrix match the row names of the input \code{countMatrix}.
}
\description{
Fits a negative binomial generalized linear mixed-effects model to each row of a provided count matrix using the \code{fitGeneralizedModel} function.
}
\seealso{
\code{\link{fitGeneralizedModel}}
}
